+++ src/core/common/message.hpp
@@ -1102,9 +1102,9 @@ public:
      *
      */
 #if OPENTHREAD_CONFIG_PLATFORM_MESSAGE_MANAGEMENT
-    uint16_t GetFreeBufferCount(void) const { return static_cast<uint16_t>(otPlatMessagePoolNumFreeBuffers()); }
+    uint16_t GetFreeBufferCount(void) const { return otPlatMessagePoolNumFreeBuffers(); }
 #else
-    uint16_t GetFreeBufferCount(void) const { return static_cast<uint16_t>(mNumFreeBuffers); }
+    uint16_t GetFreeBufferCount(void) const { return mNumFreeBuffers; }
 #endif
 
 private:
@@ -1119,9 +1119,9 @@ private:
     PriorityQueue *GetAllMessagesQueue(void) { return &mAllQueue; }
 
 #if OPENTHREAD_CONFIG_PLATFORM_MESSAGE_MANAGEMENT == 0
-    int mNumFreeBuffers;
-    Buffer mBuffers[kNumBuffers];
-    Buffer *mFreeBuffers;
+    uint16_t mNumFreeBuffers;
+    Buffer   mBuffers[kNumBuffers];
+    Buffer   *mFreeBuffers;
 #endif
     PriorityQueue mAllQueue;
 };
